added google analytics to monitor trafic.
finished workm on nadding google analytics to my website
