+++ pages/_document.js
@@ -52,29 +52,7 @@ class MyDocument extends Document {
                         href="https://fonts.googleapis.com/css2?family=Open+Sans:wght@300;500;700&display=swap"
                         rel="stylesheet"
                     />
-                    {/* Global Site Tag (gtag.js) - Google Analytics */}
-                    <Script
-                        async
-                        strategy="afterInteractive"
-                        src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GA_TRACKING_ID}`}
-                    />
-                    <Script id="google-analytics" 
-                        dangerouslySetInnerHTML={{
-                            __html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', '${process.env.NEXT_PUBLIC_GA_TRACKING_ID}', {
-              page_path: window.location.pathname,
-
-            //     <!-- this line is used to supress
-            //   the SameSite warning that will be throwed -->
-              cookie_flags: 'SameSite=None;Secure'
-            });
-          `,
-
-                        }}
-                    />
+                  
                 </Head>
                 <body>
                     <script dangerouslySetInnerHTML={{ __html: setInitialTheme }} />
